requirements.txt
numpy
pyyaml
matplotlib
# opensim -- Python bindings must be installed separately.
# See: https://opensim-org.github.io/opensim-moco-site/docs/Guides/Install/
